+++ craiyon.py
@@ -1,56 +1,5 @@
-"""
-Has the modules necessary to generate a crayion image
-"""
+from craiyon import Craiyon
 
-from os import getcwd
-from requests import post
-from json import loads
-from PIL import Image
-from io import BytesIO
-from base64 import b64decode
-from html import escape
-from sys import argv
-
-
-def sendRequest(url, headers, data):
-    resp = post(url, headers=headers, data=data)
-    return loads(resp.text)["images"]
-
-
-def image_craiyon(prompt, extension="png"):
-    """
-    """
-
-    illegalChars = '<>:"/\|?*'.join([chr(i) for i in range(32)])
-
-    #
-    prompt = escape(prompt)
-    fileFormat = extension.lower()
-
-    if fileFormat.startswith("."):
-        fileFormat = fileFormat[1:]
-
-    #
-    url = "https://backend.craiyon.com/generate"
-    headers = {"Content-Type": "application/json"}
-    data = f'"prompt": "{prompt}<br>"'
-    data = "{" + data + "}"
-
-    response = sendRequest(url, headers, data)
-
-    decoded = []
-    i = 0
-    for image in response:
-        i += 1
-        decoded.append(BytesIO(b64decode(image)))
-
-    i = 0
-    newPrompt = "creation"
-    for image in decoded:
-        i += 1
-        im = Image.open(image).resize((360, 360))
-        fileName = newPrompt + f"-{i}.{fileFormat}"
-        output = "/home/guimas/Documents/beatbot/output/" + fileName
-        im.save(output)
-
-    return output
+generator = Craiyon() # Instantiates the api wrapper
+result = generator.generate("Photorealistic image of shrek eating earth")
+result.save_images() # Saves
